% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine data sets}
\usage{
combine_data(x = list())
}
\arguments{
\item{x}{A list of data frames. All data frames must have the same number of
columns and the names of the columns must match exactly.}
}
\value{
A data table with the following variables added:
\tabular{ll}{
\code{ProjID} \tab The project ID for each individual. \cr
\code{LC_Column} \tab The LC column the sample was run on. \cr
\code{mz} \tab The m/z value calculated from the \code{Precursor mass}
and \code{Charge} variables. \cr
\code{Gene} \tab The gene, extracted from the \code{Protein description}
variable, a protein belongs to. \cr
\code{isDecoy} \tab A logical value denoting whether or not a protein
is scrambled. \cr
\code{RTmin} \tab The retention time in minutes. \cr
}
}
\description{
Combines a group of data sets output from TopPIC. This function also adds
several variables that will be used throughout the top down workflow and
removes variables that are not used.
}
